feat(database): MilkyBaseMapper 替代 BaseMapperWithCursor 并收紧逻辑删
以 MilkyBaseMapper 统一 BaseMapper 的游标遍历与 WithOptimisticLock 封装；
AbstractMpDO 使用 Long deleted 与 TableLogic，demo schema 的 deleted 列改为 bigint；
MyMetaObjectHandler 按 Long 填充 deleted；demo Mapper 改为继承 MilkyBaseMapper；
新增 MilkyBaseMapperTest；PageQueryTest 增加 @Transactional 与顺序注解避免用例间残留数据。

Made-with: Cursor

+++ demo/infrastructure/src/main/resources/sql/schema.sql
@@ -9,7 +9,7 @@ create table invocation_store
     operator_source varchar(50) not null,
     success tinyint not null,
     version int not null,
-    deleted int not null,
+    deleted bigint not null,
     gmt_create datetime not null,
     gmt_modified datetime not null,
     constraint invocation_store_id_uk_index
@@ -31,7 +31,7 @@ create table item
     store_code varchar(100) not null,
     price varchar(100),
     version int not null,
-    deleted int not null,
+    deleted bigint not null,
     channel_enum varchar(20) not null,
     ratio bigint,
     gmt_create datetime not null,
@@ -51,7 +51,7 @@ create table inventory
     amount bigint not null,
     store_code varchar(100) not null,
     version int not null,
-    deleted int not null,
+    deleted bigint not null,
     gmt_create datetime not null,
     gmt_modified datetime not null,
     constraint inventory_item_id_uk_index
@@ -71,7 +71,7 @@ create table message_store
     class_name    varchar(200) null,
     invocation_id bigint       not null,
     version int not null,
-    deleted int not null,
+    deleted bigint not null,
     gmt_create datetime not null,
     gmt_modified datetime not null,
     constraint message_store_id_uk_index
